backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-generator==1.10
attrs==25.4.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
deprecation==2.1.0
fastapi==0.128.0
fsspec==2026.1.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.0
mypy_extensions==1.1.0
outcome==1.3.0.post0
packaging==25.0
postgrest==2.27.2
propcache==0.4.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.3.1
pyroaring==1.0.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
realtime==2.27.2
requests==2.32.5
rich==14.2.0
selenium==4.40.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.50.0
storage3==2.27.2
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.2
supabase-auth==2.27.2
supabase-functions==2.27.2
tenacity==9.1.2
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==15.0.1
wsproto==1.3.2
yarl==1.22.0
zipp==3.23.0
beautifulsoup4==4.12.3
